requirements.txt
biopython==1.78
streamlit==0.71.0
numpy==1.19.1
pandas==1.1.0
python-dotenv
